Force mouse source when --forward-all-clicks
Right click and middle click require the source device to be a mouse,
not a touchscreen). Therefore, the source device was changed only when
a button other than the primary button was pressed (see
adc547fa6e8e6167cd9633a97d98de6665b8c23a).

However, this led to inconsistencies between the ACTION_DOWN when a
secondary button is pressed (with a mouse as source device) and the
matching ACTION_UP when the secondary button is released (with a
touchscreen as source device, because then there is no button pressed).

To avoid the problem in all cases, force a mouse as source device when
--forward-all-clicks is set.

Concretely, in --forward-all-clicks mode:
 - device source is set to InputDevice.SOURCE_MOUSE;
 - motion event toolType is set to MotionEvent.TOOL_TYPE_MOUSE;

For all (virtual or not) finger events:
 - device source is set to InputDevice.SOURCE_TOUCHSCREEN;
 - motion event toolType is set to MotionEvent.TOOL_TYPE_FINGER.

Fixes #3568 <https://github.com/Genymobile/scrcpy/issues/3568>

Signed-off-by: Romain Vimont <rom@rom1v.com>

TODO:
 - test that nothing is broken on a computer with a touchscreen, both
   with and without --forward-all-clicks