Use current adb port (if any) for --tcpip
If the current adb port is not 5555 (typically it is 0 because it is not
in TCP/IP mode), --tcpip automatically executes (among other commands):

    adb tcpip 5555

In case adb was already listening on another port, this command forced
to listen on 5555, and the connection should still succeed.

But this reconfiguration might be inconvenient for the user. If adb is
already in TCP/IP mode and listens on another port, use that port
without reconfiguration.

Fixes #3591 <https://github.com/Genymobile/scrcpy/issues/3591>